pyproject.toml
[tool.poetry]
name = "drf-course-work"
version = "0.1.0"
description = ""
authors = ["Anna Soboleva <asoboleva.avs@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.6"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.10"
pillow = "^11.1.0"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.4.0"
django-cors-headers = "^4.7.0"
drf-yasg = "^1.21.8"
drf-spectacular = "^0.28.0"
django-filter = "^25.1"
requests = "^2.32.3"
celery = "^5.4.0"
redis = "^5.2.1"
django-celery-beat = "^2.7.0"
coverage = "^7.6.12"


[tool.poetry.group.dev.dependencies]
pytest-cov = "^6.0.0"
pytest = "^8.3.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
